dom/media/test/test_background_video_tainted_by_drawimage.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>Test Background Video Is Tainted By drawImage</title>
<script src="/tests/SimpleTest/SimpleTest.js"></script>
<script src="manifest.js"></script>
<script src="background_video.js"></script>
<link rel="stylesheet" href="/tests/SimpleTest/test.css"/>
<script type="text/javascript">
"use strict";

var manager = new MediaTestManager;

function drawVideoToCanvas(v) {
  let w = v.width,
      h = v.height,
      c = document.createElement('canvas');
  c.width = w;
  c.height = h;
  document.body.appendChild(c);

  let gfx = c.getContext('2d');
  if (!gfx) {
    throw Error("Unable to obtain context '2d' from canvas");
  }

  gfx.drawImage(v, 0, 0, w, h);
}

startTest({
  desc: 'Test Background Video Is Tainted By drawImage',
  prefs: [
    [ "media.test.video-suspend", true ],
    [ "media.suspend-bkgnd-video.enabled", true ],
    [ "media.suspend-bkgnd-video.delay-ms", 1000 ]
  ],
  tests: gDecodeSuspendTests,
  runTest: (test, token) => {
    ok(true, `${test.name}`);
    let v = appendVideoToDoc(test.name, token);
    manager.started(token);

    waitUntilPlaying(v)
      .then(() => {
        drawVideoToCanvas(v);
        ok(v.hasSuspendTaint(), "Video is tainted after drawing to canvas");
        return checkVideoDoesntSuspend(v);
      })
      .then(() => {
        ok(true, 'Video ended before decode was suspended');
        manager.finished(token);
      })
      .catch((e) => {
        ok(false, 'Test failed: ' + e.toString());
        manager.finished(token);
      });
  }
});
</script>
